refactor: move real Span in basic-tracer package
