+++ packages/opentelemetry-basic-tracer/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './types';
+export * from './Span';
